minor4/requirements.txt
flask
flask-sqlalchemy
flask-migrate
redis
pyjwt
bcrypt
celery
psycopg2-binary